+++ src/main/java/cn/msa/msa_museum_server/entity/UserEntity.java
@@ -1,12 +1,11 @@
 package cn.msa.msa_museum_server.entity;
 
+import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
-
 import jakarta.persistence.Table;
-
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
@@ -28,5 +27,5 @@ public class UserEntity {
     public UserEntity(String username, String password) {
         this.username = username;
         this.password = password;
-	}
+    }
 }
