--- src/test/java/cn/msa/museum_server/MsaMuseumServerApplicationTests.java
@@ -1,13 +1,0 @@
-package cn.msa.museum_server;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class MsaMuseumServerApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
